# src/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: Brazilian Financial Data API
  description: |
    Unified API for accessing Brazilian financial market data from public sources.

    ## Services
    - **CVM** (port 8000): Credit market data - FIDC, FIP, FIAGRO, SECURIT
    - **BACEN** (port 8002): Central Bank data - SGS, PTAX, Focus expectations
    - **B3 CALC** (port 8001): Fixed income pricing - debentures, CRA, CRI

    ## Authentication
    All endpoints are publicly accessible (no authentication required).
  version: 2.0.0
  contact:
    name: API Support
    url: https://github.com/PedroDnT/iliquid_nightly
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server (CVM API)
  - url: http://localhost:8001
    description: Local development server (B3 CALC API)
  - url: http://localhost:8002
    description: Local development server (BACEN API)
...
paths:
...
  /api/v1/cvm/{entity}/{doc_type}:
    get:
      tags:
        - CVM
      summary: Get CVM data
      description: |
        Retrieve CVM credit market data for the specified entity and document type.

        **Entities:**
        - `fidc` - Fundos de Investimento em Direitos Creditórios
        - `fip` - Fundos de Investimento em Participações
        - `fiagro` - Fundos de Investimento nas Cadeias Agroindustriais
        - `securit` - Securitização (CRA, CRI, LCA, LCI, OTS)

        **Document Types:**
        - FIDC: mensal, cadastral, trimestral, anual
        - FIP: inf_quadrimestral, inf_trimestral, cadastral, dfin
        - FIAGRO: mensal, cadastral, trimestral, anual
        - SECURIT: cra_mensal, cri_mensal, ots_mensal, lca_mensal, lci_mensal
      operationId: getCvmData
      parameters:
        - name: entity
          in: path
          required: true
          schema:
            type: string
            enum: [fidc, fip, fiagro, securit]
          example: fidc
        - name: doc_type
          in: path
          required: true
          schema:
            type: string
          example: cadastral
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 100
          example: 100
      responses:
        "200":
          description: CVM data response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CVMDataResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /api/v1/bacen/sgs/{series_code}:
    get:
      tags:
        - BACEN
      summary: Get SGS time series
      description: Retrieve a specific time series from BACEN SGS (Sistema Gerenciador de Séries)
      operationId: getSgsSeries
      parameters:
        - name: series_code
          in: path
          required: true
          schema:
            type: integer
          example: 432
        - name: data_inicio
          in: query
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: data_fim
          in: query
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-12-31"
      responses:
        "200":
          description: Time series data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SGSSeriesResponse"
  /api/v1/bacen/sgs/multi:
    post:
      tags:
        - BACEN
      summary: Get multiple SGS series
      description: Retrieve multiple time series in a single request
      operationId: getMultiSgsSeries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultiSgsRequest"
      responses:
        "200":
          description: Multiple time series data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiSgsResponse"
  /api/v1/bacen/ptax/dolar:
    get:
      tags:
        - BACEN
      summary: Get USD/BRL exchange rate
      description: Retrieve USD to BRL exchange rate from PTAX
      operationId: getPtaxDolar
      parameters:
        - name: data
          in: query
          description: Date (YYYY-MM-DD). If not provided, returns latest.
          schema:
            type: string
            format: date
          example: "2024-01-15"
      responses:
        "200":
          description: Exchange rate data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTAXRateResponse"
  /api/v1/bacen/ptax/dolar/periodo:
    get:
      tags:
        - BACEN
      summary: Get USD/BRL exchange rate for period
      description: Retrieve USD to BRL exchange rate for a date range
      operationId: getPtaxDolarPeriod
      parameters:
        - name: data_inicio
          in: query
          required: true
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: data_fim
          in: query
          required: true
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-12-31"
      responses:
        "200":
          description: Exchange rate data for period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTAXPeriodResponse"
...
  /api/v1/bacen/ptax/moeda/{moeda}:
    get:
      tags:
        - BACEN
      summary: Get exchange rate for specific currency
      description: Retrieve exchange rate for any available currency
      operationId: getPtaxMoeda
      parameters:
        - name: moeda
          in: path
          required: true
          description: Currency code (e.g., EUR, GBP, JPY)
          schema:
            type: string
          example: EUR
        - name: data
          in: query
          description: Date (YYYY-MM-DD). If not provided, returns latest.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Exchange rate data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTAXRateResponse"
...
  /api/v1/bacen/expectativas/{endpoint_name}:
    get:
      tags:
        - BACEN
      summary: Get market expectations
      description: Query Focus/RD Mercado market expectations data
      operationId: getExpectativas
      parameters:
        - name: endpoint_name
          in: path
          required: true
          schema:
            type: string
            enum:
              - ExpectativasMercadoAnuais
              - ExpectativasMercadoMensais
              - ExpectativasMercadoTrimestrais
              - ExpectativasMercadoSelic
              - ExpectativasMercadoTop5Anuais
          example: ExpectativasMercadoMensais
        - name: indicador
          in: query
          schema:
            type: string
        - name: data
          in: query
          schema:
            type: string
            format: date
        - name: topx
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Market expectations data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpectativasResponse"
  /api/v1/bacen/taxas_juros/{endpoint_name}:
    get:
      tags:
        - BACEN
      summary: Get interest rate data
      description: Query BACEN interest rate OData endpoints
      operationId: getTaxasJuros
      parameters:
        - name: endpoint_name
          in: path
          required: true
          schema:
            type: string
          example: TaxaJurosDiaria
      responses:
        "200":
          description: Interest rate data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxaJurosResponse"
  /api/v1/b3/debentures:
    get:
      tags:
        - B3
      summary: Get debenture data
      description: Retrieve fixed income pricing for debentures
      operationId: getDebentures
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Debenture data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B3DebenturesResponse"
  /api/v1/b3/cras:
    get:
      tags:
        - B3
      summary: Get CRA data
      description: Retrieve agribusiness receivables certificates (CRA) pricing
      operationId: getCras
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: CRA data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B3CrasResponse"
  /api/v1/b3/cris:
    get:
      tags:
        - B3
      summary: Get CRI data
      description: Retrieve real estate receivables certificates (CRI) pricing
      operationId: getCris
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: CRI data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B3CrisResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: NotFound
        message:
          type: string
          example: Resource not found
        details:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
        path:
          type: string
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_previous:
          type: boolean
    CVMDataResponse:
      type: object
      properties:
        entity:
          type: string
          example: fidc
        doc_type:
          type: string
          example: cadastral
        data:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/PaginationInfo"
        metadata:
          type: object
        timestamp:
          type: string
          format: date-time
...
    SGSSeriesResponse:
      type: object
      properties:
        series_code:
          type: integer
        series_name:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              data:
                type: string
              valor:
                type: number
    MultiSgsRequest:
      type: object
      properties:
        series:
          type: array
          items:
            type: integer
        data_inicio:
          type: string
          format: date
        data_fim:
          type: string
          format: date
    MultiSgsResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SGSSeriesResponse"
    PTAXRateResponse:
      type: object
      properties:
        moeda:
          type: string
        data:
          type: string
        cotacao_venda:
          type: number
        cotacao_compra:
          type: number
    PTAXPeriodResponse:
      type: object
      properties:
        moeda:
          type: string
        data_inicio:
          type: string
        data_fim:
          type: string
        dados:
          type: array
          items:
            type: object
...
    ExpectativasResponse:
      type: object
      properties:
        endpoint:
          type: string
        data:
          type: array
          items:
            type: object
...
    TaxaJurosResponse:
      type: object
      properties:
        endpoint:
          type: string
        data:
          type: array
          items:
            type: object
    B3DebenturesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/PaginationInfo"
    B3CrasResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/PaginationInfo"
    B3CrisResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/PaginationInfo"
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NotFound
            message: "Resource not found"
            details:
              id: "123"
            timestamp: "2024-01-15T10:30:00Z"
            path: "/api/v1/cvm/fidc/mensal"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: ValidationError
            message: "Request validation failed"
            details:
              errors:
                - field: "page"
                  message: "Must be greater than 0"
            timestamp: "2024-01-15T10:30:00Z"
            path: "/api/v1/cvm/fidc/mensal"